Remove support for @lang="unspecified"
With no @lang or @lang="code_general" we do not pass on an @class to HTML and highlight.js will guess the language and try to highlight. For unhighlighted monospaced text, we use @lang="language-plaintext", which is just passed on like any language as class="langauge-plaintext". From highlight.js docs: "To apply the Highlight.js styling to plaintext without actually highlighting it, use the plaintext language"