Made test for @xml:lang more precise
As part of trouble-shooting the appearance/non-appearance of the language prefix (e.g. "[en]) in the TOC: trying to figure out why article 000658 has an [en] prefix while others do not.
Discovered that the test for @xml:lang on <title> was looking in the entire article (including bibliography) rather than just in the header. Not sure whether this will fix the problem.