changes to XSLT and CSS to support listBibl outside of back
Updated DHQ XSLT to create a template that provides a separate case for <bibl> when it appears in an Appendix. Didn't want to interfere with existing XSLT rules, but we suspect that the current provision for <bibl> in contexts outside of <back> maybe obsolete and unnecessary.

Also added a CSS rule to support the desired formatting: take the value of @label, put it at the start,  and make it bold.