Output <bibl>s instead of <biblStruct>s, save the latter elsewhere.
Added two parameters to control whether <biblStruct>s are saved in
the article or in a separate document. For now, <bibl>s are
preferable because the display stylesheets can render them.

Zotero's generated <p>-citations are parsed further, introducing
<title> in the place of a quotation mark pair, and <ref> around
URLs. This will help the XSLT do a better job matching up <bibl>s
with <biblStruct>s, though that is not yet in place.