Separate `${toDir}` property into multiple properties.
The directories for the static and proofing sites are now set in
build-properties.xml. The paths to those directories, and to the
base directory, are still set in build.xml, with ".path" added to
the property name. So, references to `${toDir.static}` may now have
been changed to `${toDir.static.path}`.

I've tested generateIssues and generateSite, and they work as
before.