Refactor article_list.xsl to use XSLT 3.0, reduce duplication.
I removed and condensed a lot of repeated code into two templates:
one that creates and sorts all the rows of the table, and one that
generates all the cells for the row corresponding to an article.

Rather than nesting a table of author names into the last cell of
each row, I set up an unordered list of names. I also removed the
"center" class from the first two cells, because the
`display:block;` CSS rule was keeping them from displaying like
regular table cells.

I also added comments and variables, and re-formatted the
whitespace for readability.