Move sorting links outside the table, into new named template.
Rather than having table headings that serve as both labels and
actionable links, I've moved the links to alternative sort pages
into a list shown before the table. `<a>` elements aren't used when
the link would be to the current page, which makes it easy to see
at a glance which sort method is applied. The display of the table
and the navigation stay consistent when navigating between pages.