Add `path_to_home` parameter to various stylesheets.
The new parameter is used to make relative links within the header
navbar and the HTML <head>. It does not yet fix links in the
sidebar navigation menu.

Because the parameter has to be set as part of the transformation,
only the regular issue indexes, contributor bios, and articles have
this working so far. But it does work!