Add internal navigation for author index.
author_index.xsl now includes a navigation bar (redesigned for
accessibility), and headings for the alphabetical groupings of
authors.

The navbar is now represented with a list inside <nav>, rather than
a table. I've adjusted the CSS to make the navbar look much like it
used to, though it will now wrap on smaller screens.