Test in TOC whether an article knows itself by the right ID.
If an article has the wrong `<idno type="DHQarticle-id">`, its link
in an issue index will be wrong (e.g.
`https://dhq-proofing.digitalhumanities.org/.html`).

It seems useful to check articles' metadata from the TOC, as the
TOC is the canonical source of truth for each issue. The two should
be in sync.